# release-21.1: storage: pool pebbleReadOnly allocations